refactor(rbac): merge test files; rename ScopeInstance → ScopeServer; add DefaultInstanceAdminPermissions
Three small follow-ups to the RBAC consolidation:

- Rename the Go const ScopeInstance → ScopeServer (the underlying KV
  string stays "instance" so this is a pure identifier rename — no data
  migration). Tygo regenerates the frontend constant.
- Merge rbac_space_test.go into rbac_test.go (~2900 lines) and
  can_space_test.go into can_test.go (~800 lines), matching the
  production-code merge from the previous round. Imports unified.
- Add DefaultInstanceAdminPermissions() for symmetry with the existing
  DefaultSpaceAdminPermissions / Moderator / Everyone helpers; switch
  InitInstanceDefaults to call it instead of inlining "all server
  permissions". Currently returns every server-scope permission, but
  carved out as its own function so future divergence (e.g. owner-only
  operations) doesn't require touching the seed loop.

Also tightened the doc comments on DefaultInstanceEveryonePermissions
and DefaultInstanceModeratorPermissions to drop the legacy "instance"
qualifier.
