+++ cli/internal/core/permission_resolver.go
@@ -24,10 +24,10 @@ import (
 // 4. If no explicit decision at current level → fall back to parent level
 //
 // This enables patterns like:
-// - #announcements rooms where "everyone" is denied message.post but
-//   "owner/admin/moderator" can still post because they have higher rank
-// - Instance admin not being blocked by a "everyone" denial because
-//   admin is checked first in the hierarchy
+//   - #announcements rooms where "everyone" is denied message.post but
+//     "owner/admin/moderator" can still post because they have higher rank
+//   - Instance admin not being blocked by a "everyone" denial because
+//     admin is checked first in the hierarchy
 //
 // The internal walk*Permission methods take a visitor callback and form the
 // single source of truth for resolution ordering. HasXxxPermission and
@@ -87,7 +87,7 @@ type visitFunc func(entry TraceEntry) visitOutcome
 // Only checks instance-level roles and KV. Used for permissions that only apply
 // at instance scope (like admin.access, space.create, dm.view).
 func (r *PermissionResolver) HasInstancePermission(ctx context.Context, userID string, perm Permission) (bool, error) {
-	if meta, known := GetPermissionMetadata(perm); known && !permissionMetadataHasScope(meta, ScopeInstance) {
+	if meta, known := GetPermissionMetadata(perm); known && !permissionMetadataHasScope(meta, ScopeServer) {
 		return false, fmt.Errorf("permission %s does not apply at instance scope", perm)
 	}
 
@@ -107,7 +107,7 @@ func (r *PermissionResolver) HasInstancePermission(ctx context.Context, userID s
 // Space-scoped permissions require the user to be a space member.
 func (r *PermissionResolver) HasSpacePermission(ctx context.Context, userID, spaceID string, perm Permission) (bool, error) {
 	if meta, known := GetPermissionMetadata(perm); known {
-		if !permissionMetadataHasScope(meta, ScopeSpace) && !permissionMetadataHasScope(meta, ScopeInstance) {
+		if !permissionMetadataHasScope(meta, ScopeSpace) && !permissionMetadataHasScope(meta, ScopeServer) {
 			return false, fmt.Errorf("permission %s does not apply at space scope", perm)
 		}
 	}
@@ -141,7 +141,7 @@ func (r *PermissionResolver) HasSpacePermission(ctx context.Context, userID, spa
 // 2. Room-level permissions: walk roles in hierarchy order, allow-or-deny per role
 // 3. Instance/space grants (fallback when no room-level decision)
 func (r *PermissionResolver) HasRoomPermission(ctx context.Context, userID, spaceID, roomID string, perm Permission) (bool, error) {
-	if !PermissionAppliesAtScope(perm, ScopeRoom) && !PermissionAppliesAtScope(perm, ScopeSpace) && !PermissionAppliesAtScope(perm, ScopeInstance) {
+	if !PermissionAppliesAtScope(perm, ScopeRoom) && !PermissionAppliesAtScope(perm, ScopeSpace) && !PermissionAppliesAtScope(perm, ScopeServer) {
 		return false, fmt.Errorf("permission %s does not apply at room scope", perm)
 	}
 
@@ -282,7 +282,7 @@ func (r *PermissionResolver) walkSpacePermission(
 	}
 
 	// Phase 2: grants in authority order (instance → space).
-	if PermissionAppliesAtScope(perm, ScopeInstance) {
+	if PermissionAppliesAtScope(perm, ScopeServer) {
 		for _, role := range instanceRoles {
 			granted, err := r.keyExists(ctx, instanceKV, rbac.AllowKey(role, parts.Verb, parts.ObjectType, rbac.ObjectIdAny))
 			if err != nil {
@@ -446,7 +446,7 @@ func (r *PermissionResolver) walkRoomPermission(
 	}
 
 	// Phase 3: instance + space grants (fallback when no room-level decision).
-	if PermissionAppliesAtScope(perm, ScopeInstance) {
+	if PermissionAppliesAtScope(perm, ScopeServer) {
 		for _, role := range instanceRoles {
 			granted, err := r.keyExists(ctx, instanceKV, rbac.AllowKey(role, parts.Verb, parts.ObjectType, rbac.ObjectIdAny))
 			if err != nil {
